requirements.txt
python-docx
customtkinter
anki
packaging
pillow